spring-data-elasticsearch-examples - moved examples to unit tests
