+++ elasticsearch/example/README.md
@@ -5,8 +5,6 @@ Requirements:
  * [Maven](http://maven.apache.org/download.cgi) - required.
  * [Elasticsearch](http://www.elasticsearch.org/overview/elkdownloads) - optional.
 
-To run example execute: ```mvn spring-boot:run```
-
 
 To use local elasticsearch cluster:
