+++ elasticsearch/example/pom.xml
@@ -7,6 +7,10 @@
     <artifactId>spring-data-elasticsearch-example</artifactId>
     <version>0.1.0</version>
 
+    <name>Spring Data Elasticsearch - Example</name>
+    <description>Sample projects for Spring Data Elasticsearch</description>
+    <url>https://github.com/spring-projects/spring-data-elasticsearch</url>
+
     <parent>
         <groupId>org.springframework.data.examples</groupId>
         <artifactId>spring-data-elasticsearch-examples</artifactId>
