--- elasticsearch/example/run-me.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-mvn spring-boot:run
