+++ elasticsearch/example/src/main/java/example/springdata/elasticsearch/repository/ConferenceRepository.java
@@ -4,7 +4,7 @@ import example.springdata.elasticsearch.model.Conference;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 /**
- * Created by akonczak on 25/10/2014.
+ * @author Artur Konczak
  */
 public interface ConferenceRepository extends ElasticsearchRepository<Conference, String> {
